Added support for moving layers up and down
Complete with updating scene and undo/redo support. However, the currently
implementation still leaves a bit of cleanup to be done. It seems we either
need some kind of wrapper class around the Map class which provides us with
all the necessary models and interfaces to make changes to the map, or we
need to put all that stuff in the Map class itself.