Added basic painting functionality to the brush
It's a start, but with a lot of missing pieces: the MapScene still needs
to be modified about tile layer changes, the user should still be able
to select which tile is used to draw and subsequent tile paints should
be merged into one undo command.

Also, it's a bad idea to hardcode the brush logic in the BrushItem. Once
another kind of brush is added, a more generic Brush interface can be
developed.