Fixed a crash on undoing object move after changing layer order
Referencing the MapObjectItem directly from MoveMapObject was a mistake,
since the item gets deleted and recreated when the scene is refreshed.
It should only reference the underlying MapObject, which is guaranteed
to be still valid.

The updating of the map object items is now done through the map
document and the map scene.