Fixed triggering endless loop in grid drawing code
When drawing a negative line with a custom dash offset, it appears Qt
can enter an endless loop. Since in these cases the grid wasn't supposed
to get drawn anyway, we can just avoid drawing negative lines.

With negative line I mean a line going from A to B, but where A is
higher than B.