Made map resizing functional
Complete with undo/redo. Objects that fall outside of the new map are
not yet deleted though, and currently it doesn't handle layers that are
not the same size as the map correctly (they are simply resized to the
new map size).

It's a bit nasty that an ObjectGroup can't be resized without having a
map owner set. This is necessary because it needs to convert tile to
pixel coordinates. Maybe I should try to represent object position using
floating point tile coordinates instead of pixels...