Fixed handling of <object> elements outside of <objectgroup>
Better not crash when this happens. Encountered it when trying to load
an old map, from when the 'objectgroup' didn't exist yet.

For backwards compatibility it might be better to implicitly create an
objectgroup instead, but I'm pretty sure I'm the only one with those
kind of maps lying around.