Added initial support for copy/paste
This implementation works by writing a TMX file to the clipboard under
the text/tmx mime type. As such it also works from one map to another or
from one Tiled instance to another.

The tricky part is that sometimes new tilesets have to be added to a map
when you paste a selection to it. Tiled tries to avoid this whenever
possible, by checking whether the same or a similar tileset (that
differs only in tile properties) is already in use. For similar
tilesets, the tile properties are merged.