Fixed an issue with the implementation of TileLayer::copy
Previously, the function could return 0 or an area smaller than the
specified region. These are confusing properties to deal with and were
the cause of errors when undoing certain stamp brush painting or cut
actions.

Also, the QRegion variant of TileLayer::copy is now no longer slower
than the QRect variant, so the latter is removed.

A memory leak was fixed in PaintTileLayer, which forgot to clean up the
copied part of the target tile layer.