Improved the support for translations
Now a regular make command will use lrelease to make .qm files from the
.ts files. Also, Tiled now looks for the translations relative to the
executable, so it is more likely to find them regardless of where it is
executed from.

It is also easier to add new translations and to update existing ones.
For the latter you can now just run 'make ts' in the 'translations'
directory, which will run lupdate with the right parameters.

qmake syntax has been largely based on the Qt Creator support for
translations.