Introduced a Preferences dialog with a few settings
The Preferences dialog currently allows the user to choose the tile
layer data format, whether to automatically reload tilesets when they
change and the interface language.

To support the latter properly, a language manager was introduced that
installs the right translators for both Tiled and Qt and figures out
the list of available translations.

This commit also adds support for saving tile layer data in the
non-binary XML format (Mantis-issue: 22).