Fixed a crash due to new size being invalid when resizing a map.
If the width or height of the map was the default value in the resize
dialog (currently 1), then the new size of the map was invalid until
the user changed the value of each axis in question.

This was because when ResizeDialog::setOldSize gets called, it set
the value of the spin boxes, causing a callback in ResizeHelper to
set the new size. However, if the width or height was default, then
the the spin box functions returned without triggering the callback,
leaving ResizeDialog::mNewSize unitialized (negative).