Fixed missing object layer properties on properties dialog
When opening the properties of an object layer from Layer -> Layer
Properties, as opposed to the layer right click menu, the object group
properties subclass was not instanciated.

I added a static function to PropertiesDialog to create the appropriate
subclass for a given layer. It's called in both places now for
consistent behavior.