Added shift modifier to bucket fill tool.
Holding shift will switch the tool to fill the current selection
regardless of the tile boundaries within it. If there is no selection,
the full layer is used.

Added callback modifiersChanged() to AbstractTool to assist tools in
updating if a keyboard modifier changes. This is achieved by installing
an event filter on the application so it works even if the map view
doesn't have focus.

Removed unnecessary calls to currentTileLayer() in
BucketFillTool::tilePositionChanged.