Further tweaks to the object properties dialog
Adjusted the margins and the resizing behaviour, as well as specifying
somewhat saner minimum and maximum values. Also made sure that the
dialog shows at its recommended size and removed an unnecessary call to
emitObjectChanged (since the undo command will do that).

Added Michael Woerister to the list of contributors.