Made copyright statements more specific
Now mentioning all contributors to each file rather than referring to
AUTHORS file. This is mainly due to this information otherwise getting
lost when parts of Tiled are used in other projects.

To avoid confusion, the first line of each header is now the file name,
to signify that the list of contributors applies to that file.

The GPL headers have also been modernized, linking to gnu.org instead of
a postal address.