Workaround a glitch when dragging a tileset tab around
QTabBar::setTabText does not check whether the tab text remains the
same, and will always relayout the tabs. This caused a glitch while
dragging a tab when the tabs don't fit horizontally.

The fact is that while dragging a tab, the tab texts will already be up
to date, so there is no need to update them.