Added import/export of tilesets to tileset context menu
The new context menu actions allow you to convert an external tileset to
an internal one and the other way around.

Unfortunately, editing of tile properties in external tilesets is still
not possible. It'll need to be temporarily converted to an internal
tileset. This is because otherwise changes to the tileset would not get
saved.