Avoid mixing debug and release dlls
By using qtLibraryTarget but no explicit separation of the resulting files,
it was easy to end up with mixed debug/release dlls, which caused Tiled to
crash on startup.

Better try to avoid the issue by using the same file name, so that a clean
build isn't mixing debug/release.