Store the tileset image width/height in the map file
Useful for when you can't load the image just at the moment, but still
want to figure out where your tiles should be.

In the future it can also be used to be able to load maps correctly
after extending the width of a tileset image, which at the moment messes
up the tile indices.