Fixed inconsistency in naming of object layers
In libtiled and in the tmx file format, they are called object groups
rather than layers. Even though it's more like a layer, this is the
traditional name, which has already spread to other libraries as well.
It makes no sense to try and change it at this point.

Not touching the UI for now.