Improved compareLayers in autoMap
compareLayers compares two tile layers.
The first tile layer is examined at QRegion r1
The second tile layer is examined at r1 + offset

When there are tiles at both places, these will be compared.
If both are unequal, the compareLayers will return false.

If there is no tile in one layer, in the other layer there must be a
tile, which must not occur in the region of the first layer.

This implementation has the advantage to need less rules at all.
Example of a corner/curve in a wall
RuleRegion is a square of 2x2:
XX
XX
RuleSet: 3 tiles are the same, one is different (lower right):
Where the X are, there should be generated Walls, where Y is, somethink else
XX
XY
When having multiple floors you had to add this Rule as well previously:
XX
XZ

With this patch it is possible to create this rule:
(empty place in lower right corner, so there can be anything except X) to be valid.
XX
X

This is one rule, but can handle multiple floors.