Added quickstamp feature
The quickstamp feature allows you to store selections with
keycombination "Ctrl + <number>".

These selections are recalled just by key "<number>".

These quickstamps are very similar to copy/paste, since they are
hold in the same datastructure, a Map*.
The difference is, that they are only usable by the specific instance of
tiled, but the quickstamps-feature has more saving-slots.

Because of the similarity, some code of "paste"-functionality was
split into another function, to make code reuse easier.