Support multiple shortcuts on certain actions
Use setShortcuts instead of setShortcut, so that all standard shortcuts
defined for a certain action get picked up. This applies for example to
cut and paste (Ctrl+X and Shift+Del now both work).

Also, Ctrl+= is now set as additional shortcut for zooming in, so that
pressing Shift is not required on US keyboard layouts.