Automap with multiple layers
In the feature Automapping, which is a rulebased support for mapping,
you can have multiple "RuleSet" layers in files defining these rules.

Actually these Layers only need to start with "RuleSet" and can continue
describing the purpose of that specific RuleSet layer.

In short: The automapping handles these multiple layers with a logic "or"
for each position.

So when the automapping compares the "RuleSet" and the "Set" layers,
it will be checked if the set layer matches any ruleset layer for
each position.