Enable/disable tools as appropriate
Now the tile-based tools are only enabled when a tile layer is selected,
and the create object tool is only enabled when an object layer is
selected.

Based on this state, the tool manager makes sure that the currently
active tool is automatically changed to an available tool. It will
prefer to select the last tool that it switched away from, to make sure
the selected tool stays the same when switching layer types and back.

To avoid confusion, AbstractTool::enable/disable have been renamed to
activate/deactivate.