Added a simple TMX Viewer application based on libtiled
It's not as minimal as I may have hoped, mainly due to still having to
write all the QGraphicsItem subclasses, but at around 200 lines of code
it's not too bad.

Customizing the graphics items and the map renderer will be necessary
anyway if you'd be writing a game that uses libtiled.