Update the current stamp when switching tilesets
A selection is kept for each tileset view, so when switching tilesets,
the tileset dock still shows the same selection that was there
previously. However, the stamp wasn't being updated to reflect the
selection in the newly selected tileset.

Updating the stamp was preferred to clearing the selection, since that
allows quickly changing between two stamps from different tilesets. The
idea is that you would rarely change tilesets anyway if you didn't have
the intention of drawing with tiles from that tileset.

Mantis-issue: 58