Fixed crash in MapDocument::unifyTilesets
unifyTilesets was initially written for MainWindow::paste, in which case
the tilesets in the map were not managed by the TilesetManager. Now it
is also used by the QuickStamp feature, which does keep references to
the tilesets in the TilesetManager.

This fix makes unifyTilesets always assume the tilesets are properly
reference counted by the TilesetManager, and modifies MainWindow::paste
to actually do that.

Two helper methods were added to TilesetManager to make adding/removing
tileset references easier.

Reviewed-by: Stefan Beller <stefanbeller@googlemail.com>