Reselect the last active map at startup
The map document, which was the current at quitting, will be activated
on startup.

Therefore the writesettings() function needs to be moved to some point
before calling confirmAllSave(). This needs to be done since
confirmAllSave messes up the last opened document. (It cycles through
all opened documents and checks if there needs a questionare for saving,
so after that function you cannot detect the document, which was the
last current document for the user.)