Fixed problems with watching tilesets multiple times
When multiple maps used the same tile set, the QFileSystemWatcher warns
about being unable to watch the file. Also, when one of those maps is
closed, the file is no longer watched even though it's still in use by
some open maps.

Fixed this by wrapping the QFileSystemWatcher with a class that deals
gracefully with files being watched multiple times.