Introduced horizontal and vertical flipping of tiles
This commit changes the TileLayer class to store Cell instances rather
than directly referring to Tile*. The Cell class contains the Tile* plus
information on whether the tile is horizontally or vertically flipped.

The MapReader and MapWriter have been adapted to store and restore the
flipped state as flags in the last bits of the global IDs (this means
older Tiled versions will refuse to load maps with flipped tiles).

The OrthogonalRenderer and IsometricRenderer have been adapted to render
the flipped tiles correctly.

This implements the first part of issue #3

Sponsored-by: Zipline Games, Inc.