Add "Execute in Terminal" option to command dialog context menu
On Linux it uses GNOME Terminal if available, otherwise xterm.
On Mac it uses a .command file to open the terminal and run the command.

When running a Cocoa app from a bundle, make sure to invoke the executable
directly instead of running with the open command if you want the output
to show up in the terminal.