Automapping: Restructure internal data structures
This replaces the translation table, which consisted of layerpointers of
rulelayers to layers within the working map, by a translation table,
which translates the layerpointers of the rulelayers into indices of the
working Map.

The set layer now is tracked by its index as well instead of a layer
pointer.

The check mechanics, if the right layer at the specified index, also changed.
That leads to dropping mAddLayers(since it's not used any more)

Advantages:
* the layer pointer changes sometimes (resizing, offsetting a map),
  so bug https://github.com/bjorn/tiled/issues/36 is fixed.

* setupMissingLayers() will check if the indices are valid, if not this
  is corrected by setting the correct index.
  If there is no such layer it is created before.
  So while the hot phase happens (automapping itself), we can rely on
  the indices, automapping is nearly the same speed (See disadvantages)

* Signals, which keep track of layer renaming, can be dismissed,
  due to the checks in setupMissingLayers() now, because the index doesn't
  change, but the internal layer pointer.

Disadvantages:
* Slightly slower (looking up by index instead
  of using the pointer directly)

The layer pointers for the rulelayers are still kept, since these are
assumed to not change.
(If the rulemap itself changes, the whole datastructures are wiped and
rebuild)

Signed-off-by: Stefan Beller <stefanbeller@googlemail.com>

remove mAddLayers