Small improvement to the Lua export plugin (incompatible)
This changes the exported Lua file to return the map table rather than
assigning it to a global variable called 'map'. This allows the map to
be loaded in Lua environments where global variables are not allowed,
and makes it easier to avoid naming conflicts between multiple maps.

OLD

  level1.lua:

    map = { ... }

  game.lua:

    require "level1"
    print(#map.layers)

NEW

  level1.lua:

    return { ... }

  game.lua

    local level1 = require "level1"
    print(#level1.layers)

Thanks to Josh Lytle for the suggestion!