libtiled: Added support for associating a polygon with a MapObject
Now a map object can optionally have a polygon associated with it, which
if specified replaces the width/height attributes of the object and
allows specifying a non-rectangular area.

Currently the polygon is always implicitly closed, but in the future
support for non-closed polygons ("polylines") is also planned.

In the TMX format, polygons are stored in the same way as in SVG.
An example:

  <object x="514" y="311">
     <polygon points="0,0 -64,32 0,96 128,64"/>
  </object>

This commit only changes libtiled to support this feature. No editing
capabilities exist yet in Tiled.

This implements the first part of issue #61

Sponsored-by: Zipline Games, Inc.