Automapping: Remove unused tile layers after applying rules
Added a cleanTileLayers() function that removes all unneccessary added
tile layers. There is cleanTilesets() which provides the same
functionality just for tilesets.

This comes handy when using rulesets such as in "The Mana World":
After all 'useful' rules, there are rules for error checking,
which i.e. covers if the automapping produces valid cave walls,
and does not break due to to small curves and abruptions.
(So you only need to check if the "error" layer is there after
having done the Automapping)

Signed-off-by: Stefan Beller <stefanbeller@googlemail.com>