Added import and export of object types
So that not every individual working on some project needs to manually
specify the colors for the different object types, but rather can import
a shared xml file defining the types and their colors.

When sometime in the future Tiled supports projects properly, the
project file would probably be a nicer place for the object type
definitions.

Sponsored-by: Zipline Games, Inc.