Center tile objects at the mouse on insertion
This fixes #67 without changing the TMX map format.

This means the bottom-left remains the origin of the tile objects even
though plain rectangle objects have their origin in the top-left. This
difference was the reason for the wrong positioning of tile objects when
inserting them, which is now fixed by taking the difference into account
in the Insert Tile Objects tool. The tool now centers the inserted tile
object at the mouse position.

The reason tile objects have their origin in the bottom-left is because
tiles in tile layers are also bottom-left aligned within their grid
cell. This in turn has its origins in top-down maps where this behavior
is usually desirable when working with tile images that are larger than
the grid size.