Improved handling of many tilesets
The tab bar is discarded for a drop down. Toolbar added for import,
export, tileset properties, and delete tileset. These entries are
removed from the context menu on TilesetView. MapDocument gains a proxy
for setting Tileset filenames and emits a signal on a change.

Handy button provided to get after tileset->setName(). Drop down is
uneditable between edit requests to assist switching the active tileset.
Tilesets are sorted alphabetically in the drop down.