Fixes to tile rotation related to oversized tiles
Tiled has specific handling of oversized tiles that was not taken into
account in the tile rotation patch. Tiles that are larger than the map
grid stick out at the top and right of a tile. This affects the
rendering code as well as the code that keeps track of the maximum tile
size used on a layer (in order to determine the area to repaint when
tiles are changed).

Also 'Left' and 'Right' are now used rather than 'CCW' and 'CW'. It's
shorter and should be clear enough.