Introduced an option to highlight the current layer
The highlighting works by darkening the layers below the current layer
and making the layers above the current layer more transparent.

This mode should help to make it clear which layer you're editing and
also serves well for checking a map for errors. The mode can be toggled
by pressing 'H'.

The degree of darkening or transparency is currently not configurable.