Added support for specifying a tile drawing offset
Each tileset can now define an offset that is applied when drawing its
tiles. This is useful when combining tilesets with a different baseline,
or lining up the base of the tiles with the tile grid.

This new tileset property is used to improve the tile positions in the
'perspective walls' and 'isometric grass and water' examples.

This closes #16

Sponsored-by: Clint Bellanger