Automapping: multiple set layers are possible now (incompatible!)
This commit will break existing rule maps, since the expected layer
naming of the rule maps changed. Instead of prefixing all layers in the
rulemap with 'rule', now the prefixes are depending on the function of
the layer.

'regions' describes where rules are placed. (Previous 'RuleRegions')
Alternatively there can be 'regions_input' and 'regions_output' which
describe the 'input_*' and the 'output_*' regions respectively.

'input<index>_<name>' replaces the 'ruleset<index>' layer.
'inputnot<index>_<name>' replaces the 'rulenotset' layer.
	The set layers name in the working map can be specified
	via the <name> part. The <index> part now seperates different
	inputs. (So input* layers having different <index> parts will belong
	to different input conditions.)
	different indexes lead to multiple conditions. Only one condition
	needs to be true and the rule gets applied. If the indexes are the
	same, then all conditions accross diffferent <name> layers must all
	be true.

'output<index>_<name>' is the replaced name of 'rule<index>_<name>'

The clearRegion function was dropped and the erase function of tilelayers
are used now.

The prepareLoad functionality was put into the constructor.
The return value of prepareLoad can be determined by checking the
length of the error string.

The functions in automapper.cpp have been reordered to fit the workflow.