Introduced string conversion functions for Map::Orientation
Almost every map reader and writer will have to deal with converting the
map orientation from and to a string. Duplicating that code makes it
cumbersome to add new orientations.

In the process removed the Hexagonal enumerator since it makes no sense
to keep that as long as there is no support for hexagonal maps yet.