Automapping: sort rules within one rulefile
As in refactoring 72e2f77ec2afbcaa0d93b6b9f0c8e56244adf8e3
the way how to obtain the ruleregions was changed to a way,
which thinks of regions instead of scanning the map for rules
line-by-line. That approach might be cleaner in theory, but the
order of rules within one rulefile is now unknown.
The original intent never was to rely on rule ordering within one
rulefile, but users did rely on it.

This commit introduces a sorting of the rules within one rulemap, so the
order is the same as it was before the rewrite of the Automapping.