Automapping: Support for Objects.
The output layers are allowed to be object layers as well.
Automapping still processes in tile coordinates, which
may yield inaccurate object handling.

For checking if objects are part of a rule or when checking
if they need to be delete first ('DeleteTiles' set to 'true'),
the object boundaries (rectangles) will be considered, which may
lead to problems, specially when using non-rectangular objects
(polygon, polyline).
Since the tile coordinates are counted in integer,
the object coordinates will be translated into these as well.
So the rectangular boundaries are rounded to nearest integer.