MapDocument::LayerType moved to Layer::Type
The move of the layer type into the layer class allows
detecting the type relying on an introduced type field of the layer
class.

Therefore asTileLayer and asObjectLayer have been implemented as
non-virtual functions in the layer class.

Additionally the functions bool isTilelayer() and bool isObjectGroup()
have been introduced for cases when you do not need the object, but only need
to know if it is a certain type.