Fixed a crash that could happen after pasting
This was a pretty nasty one to track down. At some point in time I had a
reason to copy a layer's map pointer in Layer::clone, even though the cloned
layer is not actually added to the same map. Later on I decided this was not
necessary (as can be read in the documentation), but somehow I forgot to
remove the line that copied the pointer.

The result was that the brush resulting from a paste ended up referring to a
deleted Map instance.

Fixes #130