Fixed zoom sensitivity for finer-resolution mouse wheels
This is an attempt at improving the mouse wheel zoom sensitivity for mice
with finer-resolution wheels, like Apple's Magic Mouse.

For these mice, the change in the scale is based on the current scale and the
actual wheel movement, rather than going through a fixed number of zoom
levels.

Fixes #92