Fixed object type getting erased on pressing Enter
It's not currently possible to add new object types to the model by
simply typing new names in the object properties dialog (you need to use
the preferences).

However, the default insertPolicy for QComboBox is to try adding new
strings to the model when Enter is pressed. When that failed, the value
of the combo box became empty.