Automapping: remove beforehand check if rules are used.
D. Knuth: Premature optimisation is the root of all evil.
And he is right again.

The checks if rules are used have been done before all Automapping takes
place, but a rule can alter the situation for the following rules
drastically, so either we try to correct the checks if rules are applied
or just drop these checks.

The intention of the checking before was to speed up Automapping by only
executing the needed rules, but as the checks would take more and more
code for just being correct, it might be better to remove the speedup
things completely and have a smaller code base.

Thanks to Giancarlo Todone for spotting the bug.