Automapping: Handle non existing input layers in working map.
If there is an input layer required by the rule map, which does not
exist in the working map, the rules will not get applied, as the rules
are not met.

Before there was no check if input layers existed, so you could crash
tiled.

Thanks to Giancarlo Todone for spotting the bug.

Fixes #219