Fixed resolving of symbolic links while loading map
When resolving references to external tilesets or images, it is important
to avoid resolving symbolic links. It can cause the generated relative
path when saving the map to be unnecessary complicated and possibly point
to undesired locations.

Cleaning the path should suffice for determining whether two tilesets
are the same in general.

Fixes #212