Add navigator drag&drop support
This is a quite stable running navigator or minimap.

With the help of MapRenderer a small version of the map is rendered into
a buffer image. The minimap rendering is copied from the screenshot
function and offers several flags to finetune the rendering.

For detecting mapchanges I used the indexChanged-event offered by
QUndoGroup. That does a pretty good job. Minimap rendering can be
expensive as it has to draw the entire map. To prevent slowdowns during
painting an (optional) timer is used that drops all redraw calls until
100ms are elapsed.

The overall performance impact seems to be okay. At least on my machine
(win7 64bit, quadcore i7 with good graphics card). This certainly needs
to be tested in different environments.