Optimized updating of the Tilesets dock
When switching maps the Tilesets dock was doing a lot of work, which
became noticeable with maps carrying dozens of tilesets. Two things were
done to make it fast again:

* The dock widget is temporarily hidden while it is being updated, saving
  a lot of time that was being spent needlessly relayouting the tab bar.

* The tileset models are being created on-demand, when the user actually
  switches to a certain tileset.

Now even maps with over 50 tilesets perform reasonable. If needed, it
could be further optimized by also creating the tileset views on-demand.