Automapping: Removed the no-tile exception (incompatible!)
This change is motivated by issue #306.
The no-tile exception made sense at times when the set layer was
the only input layer. (That is version 0.8.1 and before)
Now this exception causes trouble as Automapping does not
produce expected results.

This commit finally makes the Automapping incompatible to the
Automapping rules of version 0.8.1 and before.
Having spare input layers produces a different behaviour
now than before.
However not all hope is lost, the Automapping on quite filled
input layers behaves the same as before, which I expect to be the
standard usecase.

Closes #306