Performance fixes related to the Bucket Fill tool
The Bucket Fill tool would always immediately update its overlay item,
regardless of whether the tool was selected or whether the overlay was
even visible. On large maps, this caused any change of stamp to become
slow.

Another small improvement that was made was to avoid creating an overly
large fill overlay, which greatly improves the responsiveness when using
the fill tool on large maps in situations where the fill region is
small.