Implemented editing of terrain information
This is done in a dedicated dialog to avoid complicated interactions
within the main UI. The dialog can be opened from the tileset dock.

The TerrainModel now represents all terrain types of a map grouped by
the tileset they appear in.

The terrain information overlay painted on top of tiles currently only
supports tiles in orthogonal mode. It does not adjust appropriately to
isometric tiles and may also have problems with orthogonal tiles that
do not match the map grid size.

All operations are properly backed by undo/redo support.

Transition distances are now recalculated on-demand. This removes the
option of tweaking distances to achieve transition priorities. That
functionality will have to be added some other way (for example with a
separate list of transition cost modifiers).

This change also happens to close #294.