Unified image layer rendering
The image layer rendering was different for each map renderer. It seems
to me that it is better to consider the image layer as simply
independent of the tile or object layer projections, which is what this
change does.

This change also implies that image layer position is now in pixels
rather than tiles. It is currently still not possible to change this
position in Tiled, so I don't expect that to affect anybody.

Also fixed the image layer item bounds, which resolves potential repaint
issues when changing image layer opacity or visiblity.

Closes #317