Remember the size and position for some dialogs
Some dialogs were excluded from this behavior because they were either
not resizable or because it's probably better for them to always pop up
in the center (like Preferences, New Map and New Tileset).

Unfortunately for some window managers this behavior is broken and I
don't know of a good way to fix that. On Xfce for example, the dialogs
will move down by the height of their title bar each time they are
opened.

Closes #339