Depend on zlib directly when compiling against Qt 5
libtiled.pro is now set up to auto-detect a zlib installation at its
default location. When it is installed somewhere else, then the
environment variable 'ZLIB_PATH' should be set to the directory where the
'include' and 'lib' directories can be found that include zlib.

When compiling with MinGW, it will choose the static library to avoid
the need to copy a .dll around. When compiling with MSVC it will need the
zlib1.dll since I don't think the static library is compatible with that
compiler. The MSVC build has not been tested yet.

It expects zlib to be installed by using the package available at
http://gnuwin32.sourceforge.net/packages/zlib.htm

Closes #285