Implemented moving objects with arrow keys
By default the selected objects move by pixel. The Shift modifier speeds
it up to move by tile (or by subtile when fine grid snapping is
enabled). The Control modifier causes the event to be ignored, so that
it can be used to move the view.

This behavior is similar to that seen in Inkscape and GIMP.

Closes #342

Sponsored-by: Ben Wales