Changes made in property dialogs now apply immediately
This affects the custom property editing as well as the editing of any
built-in properties for classes like ObjectGroup, ImageLayer, Map and
MapObject. To make this possible, some new undo-related classes were
introduced and some new notification signals.

This change is a preparation for allowing the property dialogs to be
replaced by a docked solution.