Flare export plugin: Remove checks for certain layers being there
The flare engine evolved and supports 2 more layers optionally.
The three layers which were checked being there, should still be there,
but are not necessary. Also I cannot recall any errors due to missing
or miss spelled layers during flare development, so I think we do not
need this check for flare.

The check for these three layers unintentionally hindered other projects
using this format, which is another reason to drop the check.