Raise the Layers dock for editing a new layer's name
This made the isVisible() workaround obsolete, which had also been
copied to ObjectsDock and MapsDock where it wasn't needed in the first
place.

I decided not to automatically show the dock, since being able to edit
the new layer's name is more of a convenience that's generally not worth
popping up this dock widget for if it isn't already open.