QtPropertyBrowser: Allow controlling whether to show text for bool values
This was a rather involved patch, but it avoids the need for using
private API or duplicating QtBoolEdit (which is in a private header).

Apart from adding support for this QtBoolEdit option to the
QtCheckBoxFactory and the QtBoolPropertyManager, the setting can also
be controlled via the 'textVisible' attribute associated with
QtVariantProperty instances.