Allow changing the image layer image in the Properties dock
Introduced a 'filePath' property with a special editor that features
a small button that opens up the file chooser. It also colors the text
to make it clear whether the file path exists or not.

The VariantPropertyManager, VariantEditorFactory and FileEdit classes
were based on the example from issue 18 of the Qt Quarterly [1]

[1] http://doc.qt.digia.com/qq/qq18-propertybrowser.html