Optimized hardware-accelerated tile rendering
By using QPainter::PixmapFragment the amount of API calls to draw the
same tile over and over again is much reduced. This leads to a small
improvement in rendering speed, which is mainly noticable on maps where
large areas are filled with the same tile.

The minimum Qt version was raised to Qt 4.7 since earlier versions do
not have QPainter::PixmapFragment.