Revert "Rotate objects around their center instead of the bottom left"
This reverts commit 884184004dfa0f22a3130945cf7e86ddde01a0f9.

While rotating objects around their center may be somewhat nicer when
editing the rotation property, it has the nasty side-effect that
anything that changes the size of an object also affects its position
when this object is rotated. This means that rotated rectangles are
shifting around when you resize them, and so do polygons while editing
their nodes.

It may be possible to adjust the position to keep things on the same
place, but for now I'm opting to keep rotating objects around their
actual position.