Fixed drawing of flipped tiles when using software rendering
Due to a bug in QPaintEngineEx a negative scaleX or scaleY value results
in nothing being drawn rather than drawing a flipped image, like the
OpenGL paint engine does.

Worked this around by manually drawing the fragment and applying the
scale to the transform rather than to the target rect.