Updated the New Tileset dialog to reflect individual images option
The dialog now features a combo box where the user can choose between
creating a tileset based on a tileset image, or a tileset that is simply
a collection of images. The images still need to be added later using
the '+' button in the tileset toolbar.

The drawing offset is no longer configurable using this dialog. It can
be changed in the tileset properties instead.