Worked around an issue with static child widgets of child windows
Using static child widgets in child windows can cause a crash on exit
because there is a chance that they are still referenced after being
deleted.

In the last commit the Tile collision editor was changed to a child
window and the MapView has a static viewport widget by default. This is
flag is no longer used to avoid the above issue.